refactor: aws 관련코드 분리 ref: #14
