+++ src/main/java/com/example/jangboo/file/service/FileStorageService.java
@@ -1,0 +1,10 @@
+package com.example.jangboo.file.service;
+
+import org.springframework.web.multipart.MultipartFile;
+
+import java.io.IOException;
+
+public interface FileStorageService {
+    String uploadFile(MultipartFile file, String directory) throws IOException;
+    void deleteFile(String fileUrl);
+}
