+++ src/main/java/com/example/jangboo/infra/aws/S3Config.java
@@ -1,4 +1,4 @@
-package com.example.jangboo.global.config;
+package com.example.jangboo.infra.aws;
 
 import com.amazonaws.auth.AWSStaticCredentialsProvider;
 import com.amazonaws.auth.BasicAWSCredentials;
