+++ src/main/java/com/example/jangboo/infra/aws/S3FileStorageService.java
@@ -1,7 +1,8 @@
-package com.example.jangboo.file.service;
+package com.example.jangboo.infra.aws;
 
 import com.amazonaws.services.s3.AmazonS3Client;
 import com.amazonaws.services.s3.model.ObjectMetadata;
+import com.example.jangboo.file.service.FileStorageService;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.stereotype.Service;
 import org.springframework.web.multipart.MultipartFile;
@@ -10,16 +11,17 @@ import java.io.IOException;
 import java.util.UUID;
 
 @Service
-public class S3Service {
+public class S3FileStorageService implements FileStorageService {
 
     private final AmazonS3Client amazonS3Client;
-    public S3Service(AmazonS3Client amazonS3Client) {
+    public S3FileStorageService(AmazonS3Client amazonS3Client) {
         this.amazonS3Client = amazonS3Client;
     }
 
     @Value("${cloud.aws.s3.bucket-name}")
     private String bucket;
 
+    @Override
     public String uploadFile(MultipartFile multipartFile, String directoryPath) throws IOException {
         String originalFilename = multipartFile.getOriginalFilename();
         String uniqueFilename = UUID.randomUUID().toString().substring(0,8) + "_" + originalFilename;
@@ -33,6 +35,16 @@ public class S3Service {
 
         amazonS3Client.putObject(bucket, key, multipartFile.getInputStream(), metadata);
         return amazonS3Client.getUrl(bucket, key).toString();
+    }
+
+    @Override
+    public void deleteFile(String fileUrl) {
+        String key = extractKeyFromUrl(fileUrl);
+        amazonS3Client.deleteObject(bucket, key);
+    }
+
+    private String extractKeyFromUrl(String fileUrl) {
+        return fileUrl.substring(fileUrl.indexOf(bucket) + bucket.length() + 1);
     }
